1.2_RAG_PineCone_Streamlite/requirements.txt
streamlit
python-dotenv
langchain_huggingface
langchain_google_genai
langchain-pinecone
pymupdf
langchain-core
langchain-community
langchain